+++ app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/AddNeighbourActivity.java
@@ -77,12 +77,13 @@ public class AddNeighbourActivity extends AppCompatActivity {
             }
         });
 
+
     }
 
     @OnClick(R.id.create)
     void addNeighbour() {
         Neighbour neighbour = new Neighbour(
-                System.currentTimeMillis(),
+                (int) System.currentTimeMillis(),
                 nameInput.getEditText().getText().toString(),
                 mNeighbourImage,
                 addressInput.getEditText().getText().toString(),
